adding docs and refactoring
