+++ samples/java-customer-registry-quickstart/kafka/docker-compose.yml
@@ -19,7 +19,7 @@ services:
       - "9092:9092"
     environment:
       KAFKA_ADVERTISED_HOST_NAME: kafka
-      KAFKA_CREATE_TOPICS: "customer_updates:1:1"
+      KAFKA_CREATE_TOPICS: "customer_upserts:1:1"
       KAFKA_ZOOKEEPER_CONNECT: zookeeper:2181
     volumes:
       - /var/run/docker.sock:/var/run/docker.sock
